+++ src/enums/index.js
@@ -1,4 +1,5 @@
 
+export { HeaderMenuLinks } from './menuLinks';
 export { KeyboardKeys } from './keyboard';
 export { GeneralNoteEnums, GeneralNoteCodes } from '../components/ui/generalNote/enums';
 export { Cursors } from './styles';