+++ src/enums/menuLinks.js
@@ -1,0 +1,5 @@
+export const HeaderMenuLinks = {
+    Home: 'home',
+    Insurance: 'insurance',
+    Additional: 'additional'
+};