+++ src/layout/settings/textSize.js
@@ -4,7 +4,7 @@ import { TextSizes } from '../../settings';
 import { ButtonTypes } from '../../components/common/ui';
 import DataStorage from '../../services/storage';
 import { SaveObjects } from '../../components/common/saves';
-import { IdService } from '../../services';
+import { IdService, StyleService } from '../../services';
 
 class SettignsTextSize extends HTMLElement {
     constructor() {
@@ -56,7 +56,7 @@ class SettignsTextSize extends HTMLElement {
         const body = document.querySelector('body');
         const indicator = IdService.id('indicator', this.shadow);
         indicator.innerHTML = `<b>${size}</b>`;
-        body.style.fontSize = `${size}%`;
+        StyleService.setProperty(body, 'fontSize', size+'%');
     }
 
     render() {