+++ src/pages/start/theme.js
@@ -1,0 +1,34 @@
+import { Themes } from '../../theme/enums';
+
+export const StartPageTheme = {
+    [Themes.main1]: {
+        startItem: {
+           border: '#dcdcdc',
+           label: '#bf2685'
+        }
+    },
+    [Themes.black]: {
+        startItem: {
+            border: '#dcdcdc',
+            label: '#282929'
+        }
+    },
+    [Themes.red]: {
+        startItem: {
+            border: '#dcdcdc',
+            label: '#c21b1b'
+        }
+    },
+    [Themes.blue]: {
+        startItem: {
+            border: '#dcdcdc',
+            label: '#2c9cd4'
+        }
+    },
+    [Themes.yellow]: {
+        startItem: {
+            border: '#dcdcdc',
+            label: '#e3ac0b'
+        }
+    },
+};