+++ src/services/utils/index.js
@@ -1,6 +1,6 @@
 export { simulateDelay } from './server';
 export { showComponent } from './render';
-export { getVersionFromPackage } from './strings';
+export { getVersionFromPackage, capFirstLetter } from './strings';
 export { randomInteger, sample } from './number';
 export { isMobile } from './mobile';
 export { objectPropertyAmount } from './object';