+++ src/services/utils/strings.js
@@ -6,3 +6,7 @@ export const getVersionFromPackage = (version) => {
     ver = [ver.slice(0, position), dot, ver.slice(position)].join('');
     return ver;
 }
+
+export const capFirstLetter = (str) => {
+    return str.charAt(0).toUpperCase()+ str.slice(1);
+}