+++ Changelog.md
@@ -1,6 +1,7 @@
 ## Changelog
 v 2.48
-  Added improvements to Welcome page, Added Features page
+  Added improvements to Welcome page, Added Features page, 
+  Style setProperty implementation
 
 v 2.46
   Created Start intial version, Added Welcome page, Improved Tabs structure